Move build info command to dedicated command file
