+++ GPTifier/include/commands.hpp
@@ -1,3 +1,4 @@
 #pragma once
 
+void print_build_information();
 void create_chat_completion();
